Move messenger/ to chat_service/
