+++ parlai/chat_service/tasks/overworld_demo/config.yml
@@ -15,7 +15,7 @@ tasks:
     timeout: 180
     agents_required: 2
 task_name: overworld_demo
-world_module: parlai.messenger.tasks.overworld_demo.worlds
+world_module: parlai.chat_service.tasks.overworld_demo.worlds
 overworld: MessengerOverworld
 page_id: 1 # Configure Your Own Page
 max_workers: 30
