+++ parlai/chat_service/tasks/qa_data_collection/config.yml
@@ -5,7 +5,7 @@ tasks:
     timeout: 300
     agents_required: 1
 task_name: qa_data_collection
-world_module: parlai.messenger.tasks.qa_data_collection.worlds
+world_module: parlai.chat_service.tasks.qa_data_collection.worlds
 overworld: MessengerOverworld
 page_id: 1 # Configure Your Own Page
 max_workers: 30
