+++ parlai/chat_service/tasks/qa_data_collection/worlds.py
@@ -3,7 +3,7 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-from parlai.messenger.core.worlds import (
+from parlai.chat_service.services.worlds import (
     SimpleMessengerOverworld as MessengerOverworld,
     OnboardWorld as QADataCollectionOnboardWorld,
 )
